% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterToBSGenome.R
\name{characterToBSGenome}
\alias{characterToBSGenome}
\title{characterToBSGenome}
\usage{
characterToBSGenome(genome.name)
}
\arguments{
\item{genome.name}{a character string uniquely identifying a \code{\link{BSgenome}} (e.g. "hg19", "mm10" are ok, but "hg" is not)}
}
\value{
A \code{\link{BSgenome}} object
}
\description{
Given a character string with the "name" of a genome, it returns a \code{\link{BSgenome}} object if available.
}
\note{
This function is memoised (cached) using the \code{memoise} package. To empty the cache, use \code{forget(charecterToBSGenome)}
}
\examples{
g <- characterToBSGenome("hg19")
}
\seealso{
\code{\link{getGenomeAndMask}}, \code{\link{maskFromBSGenome}}
}
