% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringScore.R
\name{clusteringScore}
\alias{clusteringScore}
\title{Clustering Score}
\usage{
clusteringScore<-function(A,nt=100,neighbors=3,...)
}
\arguments{
\item{A}{the original set of regions to get a similar one. A region set in any of the accepted formats by \code{\link{toGRanges}} (\code{\link{GenomicRanges}}, \code{\link{data.frame}}, etc...)}

\item{nt}{numeric value corresponding to the number of permutations.}

\item{neighbors}{numeric value corresponding to the number of sets considered to calculate the mean distance..}
}
\value{
It returns a list of class \code{permTestResults}.
}
\description{
Given a set of regions A and a number of neighbor, returns a p-value
that shows how clustered is this set.
}
\details{
This function calculates the distance to n neighbors for each set
and obtain the mean of this distances. The p-value indicates how 
probable is to obtain this mean for a random set.
}
\examples{
A <- createRandomRegions(nregions=100, length.mean=500, length.sd=100)
p <- clusteringScore(A)

}
\seealso{
\code{\link{permTest}}
}
