% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emptyCacheRegioneR.R
\name{emptyCacheRegioneR}
\alias{emptyCacheRegioneR}
\title{Empty Cache regioneR}
\usage{
emptyCacheRegioneR()
}
\value{
The cache is emptied
}
\description{
Empties the caches used by the memoised function in the regioneR package.
}
\examples{
emptyCacheRegioneR()
}

