% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromosomesByOrganism.R
\name{getChromosomesByOrganism}
\alias{getChromosomesByOrganism}
\title{getChromosomesByOrganism}
\usage{
getChromosomesByOrganism()
}
\value{
a list with the organism as keys and the list of available chromosome sets as values
}
\description{
Function to obtain a list of organisms with their canonical and (when applicable) the autosomal chromosome names.
This function is not usually used by the end user directly but through the filterChromosomes function.
}
\examples{

chrsByOrg <- getChromosomesByOrganism() 
chrsByOrg[["hg"]]
chrsByOrg[["hg"]][["autosomal"]]

}
\seealso{
\code{\link{getGenome}}, \code{\link{filterChromosomes}}
}
