% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskFromBSGenome.R
\name{maskFromBSGenome}
\alias{maskFromBSGenome}
\title{maskFromBSGenome}
\usage{
maskFromBSGenome(bsgenome)
}
\arguments{
\item{bsgenome}{a \code{\link{BSgenome}} object}
}
\value{
A \code{\link{GRanges}} object with the active mask in the \code{\link{BSgenome}}
}
\description{
Extracts the merge of all the active masks from a \code{\link{BSgenome}}
}
\note{
This function is memoised (cached) using the \code{\link{memoise}} package. To empty the cache, use \code{\link{forget}(maskFromBSGenome)}
}
\examples{
g <- characterToBSGenome("hg19")

maskFromBSGenome(g)

}
\seealso{
\code{\link{getGenomeAndMask}}, \code{\link{characterToBSGenome}}, \code{\link{emptyCacheRegioneR}}
}

