% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarRegions.R
\name{similarRegions}
\alias{similarRegions}
\title{Similar Regions}
\usage{
similarRegions<-function(A,genome="hg19",k=0.005,iter=1000,evaluate.function,...)
}
\arguments{
\item{A}{the original set of regions to get a similar one. A region set in any of the accepted formats by \code{\link{toGRanges}} (\code{\link{GenomicRanges}}, \code{\link{data.frame}}, etc...)}

\item{genome}{the reference genome to use. A valid genome object.}

\item{k}{numeric value corresponding to the tolerance parameter.}

\item{iter}{numeric value corresponding to the maximum number of iterations.}

\item{evaluate.function}{function to search for association. It must return a numeric value.}
}
\value{
It returns a \code{\link{GenomicRanges}} object with the regions obtained.
}
\description{
Given a set of regions A and a evaluation function, returns another set of
regions with  a similar evaluation according to this evaluation function.
}
\details{
A random set of regions will be created with similar scores according the evaluation function,
maintaining the length mean and standard deviation of the original set.
}
\examples{
A <- createRandomRegions(nregions=20, length.mean=1000, length.sd=200)
B <- similarRegions(A=A,evaluate.function=width)

}
\seealso{
\code{\link{toGRanges}}, \code{\link{createRandomRegions}}
}
