% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.permTestResults.R
\name{summary.permTestResults}
\alias{summary.permTestResults}
\title{Summary of permTestResults objects}
\usage{
\method{summary}{permTestResults}(object, ...)
}
\value{
the summary is printed
}
\description{
Summary of permTestResults objects
}
\keyword{internal}

