% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.permTestResultsList.R
\name{summary.permTestResultsList}
\alias{summary.permTestResultsList}
\title{Summary of permTestResultsList objects}
\usage{
\method{summary}{permTestResultsList}(object, ...)
}
\value{
the summary is printed
}
\description{
Summary of permTestResultsList objects
}
\keyword{internal}

